feat(ui): add mini/advanced controlnet ui
